Cap offset at min of 0

+++ src/main/java/dev/yurisuika/raised/server/command/RaisedCommand.java
@@ -38,7 +38,7 @@ public class RaisedCommand {
                                     context.getSource().sendFeedback(new TranslatableText("commands.raised.value.hud.query", config.value.hud), false);
                                     return 1;
                                 })
-                                .then(argument("value", IntegerArgumentType.integer())
+                                .then(argument("value", IntegerArgumentType.integer(0))
                                         .executes(context -> {
                                             setHud(IntegerArgumentType.getInteger(context, "value"));
                                             context.getSource().sendFeedback(new TranslatableText("commands.raised.value.hud.set", config.value.hud), false);
@@ -51,7 +51,7 @@ public class RaisedCommand {
                                     context.getSource().sendFeedback(new TranslatableText("commands.raised.value.chat.query", config.value.chat), false);
                                     return 1;
                                 })
-                                .then(argument("value", IntegerArgumentType.integer())
+                                .then(argument("value", IntegerArgumentType.integer(0))
                                         .executes(context -> {
                                             setChat(IntegerArgumentType.getInteger(context, "value"));
                                             context.getSource().sendFeedback(new TranslatableText("commands.raised.value.chat.set", config.value.chat), false);
